Upgrades and mild refactoring
