+++ package.json
@@ -18,23 +18,24 @@
   },
   "devDependencies": {
     "@types/ini": "^1.3.29",
-    "@types/jest": "^25.1.4",
+    "@types/jest": "^26.0.22",
     "@types/js-yaml": "^3.5.29",
     "@types/node": "^6.0.52",
-    "@types/ramda": "^0.26.44",
+    "@types/ramda": "^0.27.39",
     "husky": "^4.2.3",
-    "jest": "^25.1.0",
+    "jest": "^26.6.3",
     "lint-staged": "^10.0.8",
-    "prettier": "^1.12.1",
-    "ts-jest": "^25.2.1",
-    "typescript": "^3.4"
+    "prettier": "^2.2.1",
+    "ts-jest": "^26.5.4",
+    "typescript": "^4.2.3"
   },
   "scripts": {
     "build": "yarn tsc",
     "in-rebase": "(! test -d '$(git rev-parse --git-path rebase-merge)' && ! test -d '$(git rev-parse --git-path rebase-apply)')",
     "lint": "yarn tsc --noEmit",
     "prepublish": "yarn build",
-    "test": "yarn lint && yarn jest"
+    "test": "yarn lint && yarn jest",
+    "test-no-lint": "yarn jest"
   },
   "lint-staged": {
     "src/**/*.{js,jsx,ts,tsx}": [
@@ -53,9 +54,6 @@
       "pre-commit": "yarn lint ; (yarn in-rebase && lint-staged)",
       "pre-push": "CI=1 yarn test"
     }
-  },
-  "resolutions": {
-    "acorn-globals/acorn": "^6.4.1"
   },
   "sideEffects": false
 }
